+++ tests/Sniffs/CompositeCodeElementSniffTest.php
@@ -22,6 +22,7 @@ class CompositeCodeElementSniffTest extends TestCase
         $dataSets[] = [
             [
                 'FqcnMethodSniff.enabled' => 'true',
+                'FqcnMethodSniff.reportMissingTags' => true,
             ],
             __DIR__ . '/fixtures/TestClass0.php',
             [
@@ -54,6 +55,7 @@ class CompositeCodeElementSniffTest extends TestCase
         $dataSets[] = [
             [
                 'FqcnMethodSniff.invalidTags' => ['@SmartTemplate'],
+                'FqcnMethodSniff.reportMissingTags' => true,
             ],
             __DIR__ . '/fixtures/TestClass1.php',
             [
@@ -90,6 +92,7 @@ class CompositeCodeElementSniffTest extends TestCase
         $dataSets[] = [
             [
                 'FqcnMethodSniff.enabled' => 'false',
+                'FqcnMethodSniff.reportMissingTags' => true,
             ],
             __DIR__ . '/fixtures/TestClass1.php',
             [
@@ -102,6 +105,7 @@ class CompositeCodeElementSniffTest extends TestCase
         $dataSets[] = [
             [
                 'useReflection' => true,
+                'FqcnMethodSniff.reportMissingTags' => true,
             ],
             __DIR__ . '/fixtures/TestClass3.php',
             [
@@ -120,6 +124,7 @@ class CompositeCodeElementSniffTest extends TestCase
         $dataSets[] = [
             [
                 'useReflection' => false,
+                'FqcnMethodSniff.reportMissingTags' => true,
             ],
             __DIR__ . '/fixtures/TestClass4.php',
             [
@@ -138,6 +143,7 @@ class CompositeCodeElementSniffTest extends TestCase
         $dataSets[] = [
             [
                 'useReflection' => false,
+                'FqcnMethodSniff.reportMissingTags' => true,
             ],
             __DIR__ . '/fixtures/TestClass5.php',
             [
@@ -191,6 +197,7 @@ class CompositeCodeElementSniffTest extends TestCase
         $dataSets[] = [
             [
                 'useReflection' => false,
+                'FqcnMethodSniff.reportMissingTags' => true,
             ],
             __DIR__ . '/fixtures/TestTrait7.php',
             [
@@ -203,6 +210,7 @@ class CompositeCodeElementSniffTest extends TestCase
         $dataSets[] = [
             [
                 'useReflection' => false,
+                'FqcnMethodSniff.reportMissingTags' => true,
             ],
             __DIR__ . '/fixtures/TestInterface8.php',
             [
@@ -214,6 +222,7 @@ class CompositeCodeElementSniffTest extends TestCase
         $dataSets[] = [
             [
                 'useReflection' => false,
+                'FqcnMethodSniff.reportMissingTags' => true,
             ],
             __DIR__ . '/fixtures/TestClass7.php',
             [
@@ -225,6 +234,7 @@ class CompositeCodeElementSniffTest extends TestCase
         $dataSets[] = [
             [
                 'useReflection' => false,
+                'FqcnMethodSniff.reportMissingTags' => true,
             ],
             __DIR__ . '/fixtures/TestClass8.php',
             [
@@ -247,7 +257,8 @@ class CompositeCodeElementSniffTest extends TestCase
         $dataSets[] = [
             [
                 'useReflection' => false,
-                'addViolationId' => true
+                'addViolationId' => true,
+                'FqcnMethodSniff.reportMissingTags' => true,
             ],
             __DIR__ . '/fixtures/TestClass5.php',
             [
@@ -282,6 +293,7 @@ class CompositeCodeElementSniffTest extends TestCase
         $dataSets[] = [
             [
                 'useReflection'                     => false,
+                'FqcnMethodSniff.reportMissingTags' => true,
             ],
             __DIR__ . '/fixtures/TestClass9.php',
             [
@@ -309,7 +321,8 @@ class CompositeCodeElementSniffTest extends TestCase
         $dataSets[] = [
             [
                 'useReflection' => false,
-                'reportType' => 'error'
+                'reportType' => 'error',
+                'FqcnMethodSniff.reportMissingTags' => true,
             ],
             __DIR__ . '/fixtures/TestInterface8.php',
             [], // 1 error reported, 0 warnings